Allow basic auth on the /metrics endpoint
