+++ crates/maker/src/routes.rs
@@ -29,6 +29,7 @@ use rocket::response::Responder;
 use rocket::serde::json::Json;
 use rocket::State;
 use rocket_cookie_auth::auth::Auth;
+use rocket_cookie_auth::basic::BasicAuthGuard;
 use rocket_cookie_auth::forms::ChangePassword;
 use rocket_cookie_auth::forms::Login;
 use rocket_cookie_auth::user::User;
@@ -320,11 +321,9 @@ pub async fn get_cfds<'r>(
     }
 }
 
-// TODO: Use non-cookie auth for /metrics endpoint as Prometheus does not
-// support cookie-auth (for now, leave unauthenticated)
 #[rocket::get("/metrics")]
 #[instrument(name = "GET /metrics", skip_all, err)]
-pub async fn get_metrics<'r>() -> Result<String, HttpApiProblem> {
+pub async fn get_metrics<'r>(_basic_auth: BasicAuthGuard) -> Result<String, HttpApiProblem> {
     let metrics = prometheus::TextEncoder::new()
         .encode_to_string(&prometheus::gather())
         .map_err(|e| {
