--- crates/rocket-basicauth/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "rocket-basicauth"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-base64 = "0.13"
-hex = "0.4"
-rocket = { version = "0.5.0-rc.2", features = ["json"] }
-void = "1"
