+++ crates/rocket-cookie-auth/src/lib.rs
@@ -5,6 +5,7 @@ use anyhow::Result;
 use rocket::async_trait;
 
 pub mod auth;
+pub mod basic;
 pub mod error;
 pub mod forms;
 mod session;
